+++ threedi_model_migration/api_utils.py
@@ -25,27 +25,47 @@ import hashlib
 import json
 import logging
 import time
+from enum import Enum
 
 
 logger = logging.getLogger(__name__)
 
 
+class PushMode(Enum):
+    full = "full"  # push all revisions; error if API has a different commit with same number
+    overwrite = "overwrite"  # always delete the API schematisation (for testing mostly)
+    incremental = (
+        "incremental"  # push newer revisions, increase revision number if necessary
+    )
+    never = "never"
+
+
+class NoSchematisation(Exception):
+    pass
+
+
 def get_or_create_schematisation(
-    api: V3BetaApi, schematisation: Schematisation, overwrite: bool = False
+    api: V3BetaApi, schematisation: Schematisation, mode: PushMode = PushMode.full,
 ) -> OASchematisation:
+    if mode is PushMode.never:
+        raise ValueError("Invalid push mode 'never'")
     resp = api.schematisations_list(
         slug=schematisation.slug, owner__unique_id=schematisation.metadata.owner
     )
-    if resp.count == 1 and not overwrite:
+    if resp.count == 1 and mode in {PushMode.full, PushMode.incremental}:
         logger.info(
             f"Schematisation '{schematisation.slug}' already exists, skipping creation."
         )
         return resp.results[0], False
-    elif resp.count == 1 and overwrite:
+    elif resp.count == 1 and mode is PushMode.overwrite:
         logger.info(
             f"Schematisation '{schematisation.name}' already exists, deleting..."
         )
         delete_schematisation(api, resp.results[0].id)
+    elif resp.count == 0 and mode is PushMode.incremental:
+        raise NoSchematisation(
+            "Cannot incrementally update '{schematisation.slug}' as it does not exist."
+        )
 
     logger.info(f"Creating schematisation '{schematisation.slug}'...")
     obj = OASchematisation(