+++ CHANGES.rst
@@ -9,6 +9,10 @@ History
 
 - Added an amqp consumer.
 
+- Fixed bug in sqlite zipping.
+
+- Added sentry support.
+
 
 0.4.0 (2021-12-08)
 ------------------